Implement more tests, general minor fixes. Module importing issue for pytest persists, add app. before each local module

+++ app/main.py
@@ -3,16 +3,20 @@ from fastapi import FastAPI, APIRouter, Query, Path, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from typing import List, Dict, Annotated
 from uuid import UUID
+# from database import init_db
+# from models import UserInDB
+# from schemas import UserCreate, UserOut, UserUpdate, Role
+# from utils import generate_users
 
-from database import init_db
-from models import UserInDB
-from schemas import UserCreate, UserOut, UserUpdate, Role
-from utils import generate_users
+from app.database import init_db
+from app.models import UserInDB
+from app.schemas import UserCreate, UserOut, UserUpdate, Role
+from app.utils import generate_users
 
 app = FastAPI()
 router = APIRouter()
 origins = [
-    "http://localhost:4200",
+    "*",
 ]
 
 app.add_middleware(
@@ -37,12 +41,14 @@ async def get_all_users() -> List[UserOut]:
         users_cursor = UserInDB.find()
         users = await users_cursor.to_list(length=None)
     except Exception as e:
-        raise HTTPException(status_code=500,
-                            detail=f"An error occured while fetching users from the database. Please try again.{e}")
+        raise HTTPException(
+            status_code=500,
+            detail=f"An error occurred while fetching users from the database. Please try again.{e}"
+        )
 
     users_out_list = []
     for user in users:
-        print(user.id)
+        # print(user.id)
         users_out_list.append(UserOut(**user.dict()))
     return users_out_list
 
@@ -51,19 +57,23 @@ async def get_all_users() -> List[UserOut]:
 async def create_user(user: UserCreate) -> UserOut:
     """ Create a user. Returns the user data from the database after saving said user, including the newly
     created ID."""
-
+    
     try:
         user_in_db = UserInDB(**user.dict())
         db_user = await user_in_db.insert()
     except Exception as e:
-        raise HTTPException(status_code=500,
-                            detail="An error occured while persisting the user to the database. Please try again.")
+        raise HTTPException(
+            status_code=500,
+            detail="An error occurred while persisting the user to the database. Please try again."
+        )
     user_out_data = db_user.dict(exclude={'id'})
     user_out = UserOut(id=str(db_user.id), **user_out_data)
     if user_out:
         return user_out
     else:
-        raise HTTPException(status_code=500, detail="Could not create user. Please try again.")
+        raise HTTPException(
+            status_code=500, detail="Could not create user. Please try again."
+        )
 
 
 @router.delete("/users/{doc_id}", status_code=200)
@@ -82,15 +92,19 @@ async def delete_user(doc_id: Annotated[UUID, Path(title="The ID of the user tha
 @router.delete("/users", status_code=200)
 async def delete_all_users():
     all_users = await UserInDB.find_all().to_list()
-
     # Delete each user one by one (optional)
-    for user in all_users:
-        await user.delete()
+    if all_users:
+        for user in all_users:
+            await user.delete()
 
-    # Alternatively, you can delete all users in a single operation (recommended)
-    # await UserInDB.delete_many({})
+        # Alternatively, you can delete all users in a single operation (recommended)
+        # await UserInDB.delete_many({})
 
-    return {"message": "All users have been deleted."}
+        return {"message": "All users have been deleted."}
+    raise HTTPException(
+        status_code=404,
+        detail="No users were found."
+    )
 
 
 # GET USERS BY EMAIL
@@ -101,7 +115,7 @@ async def get_by_email(email: Annotated[str, Query(title="The email of the user 
         user = await UserInDB.find_one(UserInDB.email == email)
     except Exception as e:
         raise HTTPException(status_code=500,
-                            detail="An error occured while fetching the data from the database. Please try again.")
+                            detail="An error occurred while fetching the data from the database. Please try again.")
     if not user:
         raise HTTPException(status_code=404, detail=f"user with email: {email} not found.")
     # iterate through the list and create a UserOut object for each entry
@@ -124,7 +138,7 @@ async def get_user_by_id(
         user_out = UserOut(**user.dict())
     except Exception as e:
         raise HTTPException(status_code=500,
-                            detail="An error occured while fetching the data from the database. Please try again.")
+                            detail="An error occurred while fetching the data from the database. Please try again.")
     return user_out
 
 
@@ -133,14 +147,13 @@ async def populate_db(
         count: Annotated[int, Query(title="Count represents the number of entries the DB will be populated with.")]
 ) -> List[UserOut]:
     """Populate the database with dummy data. Through Query parameters the user can decide the amount
-     of users to populate the db with."""
+    of users to populate the db with."""
     users_create = generate_users(count)
     users_out = []
     for user in users_create:
         db_user = await UserInDB(**user.dict()).insert()
         user_out_data = db_user.dict(exclude={'id'})
         user_out = UserOut(id=str(db_user.id), **user_out_data)
-        print(user_out.id)
         users_out.append(user_out)
     return users_out
